requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
passlib==1.7.4
psycopg2==2.9.3
pydantic==1.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.18.2
